backend/requirements.txt
fastapi
uvicorn
requests
python-multipart
openai-whisper
torch
